SeqExtStepPage: replace float fov_pixels_per_tick with Q8 integer, saving ~400 bytes flash
16385 → Q8 (×256) fixed-point uint16_t for fov_pixels_per_tick field,
eliminating ~15 float operations (__mulsf3, __divsf3, ceilf, float
conversions) in the pianoroll rendering path.

+++ src/mcl/MCL/SeqExtStepPage.h
@@ -36,7 +36,7 @@ public:
 
   static constexpr uint8_t zoom_max = 32;
 
-  float fov_pixels_per_tick;
+  uint16_t fov_pixels_per_tick; // Q8: actual_value × 256
 
   int16_t cur_x;
   int16_t cur_y;
